feat: add APCS page
- Add APCS page and new componnet - Teacher Card
