--- src/pages/apcs/data.tsx
@@ -1,26 +1,0 @@
-import { TeacherCardProps } from "@/components/TeacherCard";
-
-export const teachersCardData: TeacherCardProps[] = [
-  {
-    name: "Nobel",
-    imageUrl: "avatar_1.png",
-    major: "清大資工博士",
-    experience: ["10年程式家教與團班講師", "清大資工、電機教師"],
-  },
-  {
-    name: "Johnson",
-    imageUrl: "avatar_2.png",
-    major: "交大資工博士",
-    experience: ["10年程式家教與團班講師", "外商公司 面試官"],
-  },
-  {
-    name: "Michael",
-    imageUrl: "avatar_6.png",
-    major: "清大資工博士",
-    experience: [
-      "7年程式家教與團班講師",
-      "2017 APCS 觀念+實作接近滿分",
-      "2023-2024 共指導8位學生APCS總級分7分以上",
-    ],
-  },
-];
